Fix for CR 2005; reviewed by tpot
Adding caching of NET_USER_INFO_3 after a successful sam_logon
request.  This means that the query_user() and lookup_usergroups()
will succeed in a Win2k domain with RA==1 environment even without
an --auth-user defined in secrets.tdb.

Introduced new samlogon_cache.tdb used only by winbindd at this point.

This is a APP_HEAD only change for now.  I need to consider if and how
to port it to HEAD/SAMBA_3_0.