updated the TCON test so that win2000 passes. Samba now fails this test.
It is an interesting test because it shows that win2000 completely
ignores the TID and VUID fields in a SMBwriteX. In Samba it is hard to
do this as we check the VUID and TID fields before we call the SMB
specific reply functions.

The test also shows that the list of open files must be global to the
socket, not specific to a tcon.
(This used to be commit be98069c4e5bbfbe3ce66c20f796c2d2324e7511)