Don't try and dlsym or dlclose a NULL pointer.
The new modules system does not always dlopen() it's modules, and when it
does, it keeps them open for the life of the server, not the life of the
connection.

This caused a segfault on every tree disconnect!

Andrew Bartlett
(This used to be commit c76ecbae6295022d031d2e286f2d67e5d08946a2)