Fix a botched merge from appliance-head...
We have an extra NULL here, which is (if I understand correctly) related
to a optional 'charset conversion' function.

In any case, as this is a variable args length function, the compiler
didn't pick it up.

Would it be possible to change applicance_head (or even HEAD) in such
a way that this extra arg would cause a warning?

Finding this by testing is getting painful...

Andrew Bartlett