The code is a bit of a mess right now.
Added code to add keys to the in-momory key tree.
Added code to convert SID strings to SIDs in memory format.
Added code to construct initial security descriptors and to inherit
security descriptors.

A couple of more hours should see the code to add keys and values complete.

Then I have to start on the code to write out a registry :-(
(This used to be commit fc691a9a5e86686ebd0b3bab75a286488ab6ca57)