Well, we can now add keys and values and delete keys and values ...
Lots more to do, like handling regular expressions for keys, and input of
the security descriptors, etc. However, I now have to work on writing the
in-memory tree out to a file in a format that Windows can understand.
(This used to be commit 562b36510c704eba84b036aacdb1f984339afafc)