Now that Volker fixed the real issues with ldapsam and adding null attributes etc, move the SAMR create_user code back to using the 'pdb_init_sam_pw' method to fill out the attributes.
This is basicly the same code, but we really didn't need the duplication.

Also, take advantage of the fact that RIDs will always be returned back into
the SAM_ACCOUNT on ADD, so we don't need to duplicate the 'get'.

This should also help in sites with replicated LDAP - the second fetch might
occour before the first is replicated back.

Andrew Bartlett
(This used to be commit 39714c24fd9da4701d4fe69ddd3d61a25254409f)