As nobody really objected to this patch, I opted to create facts :-)
This uses 'socket address' as the source address for nmbd. This way we
can again synchronize with the DMB if we have 'bind interfaces only'
to a virtual interface.

I'd love to see this in 2.2.9, but that is up to jerry or jra.

Volker
(This used to be commit fe637c690b671ddb5ccbf506825a9ede6abf6668)