Merge Samba 3.0 pdb_ldap from 3.0 into HEAD, so as to allow idra to continue his IDMAP work.
This version also works properly (the HEAD version had suffered from bitrot),
and should be a good basis to change into the new IDMAP rules.

It also includes UTF8 conversions.

Included also are the schema changes, and a note about the now very old scripts
in examples/LDAP (they don't work for this, or even the previous schema).

Andrew Bartlett