configure.in updates:
Now the IRIX and non-irix cases for one of these switch statements is the same,
eliminate the statement...

We now use autoconf > 2.50, so we can use some of it's features.

We also need to correctly include the magic for building vfs_fake_perms.
(This used to be commit a4ec8a615195f681f315ce23e799f2e7d35f27f6)