SMB Signing with NTLMv2 works!
(well, under certain conditions :-)

There is no length limit on the size of the authentication response added
into the MD5 hash.  (We had previously limited this to lengths like 40, 44 or
64 in attempts to make sense of what the SNIA spec tells us).

Instead, the entire authentication response is added in.

Currently, this only works on a Win2k domain members with a Samba PDC,
becouse our NTLMv2 code currently fails against an Win2k PDC.

However, this splits the problem in half - particularly as the NTLMv2 format
is known, and even has an ethereal disector! (thanks tpot).

Andrew Bartlett