i'm getting rather sick of this
**********************************************************
**** PLEASE TEST YOUR CHANGES BEFORE CHECKING THEM IN ****
**********************************************************

Jelmer, your change did not fix the -b option, it just make it not
segfault on debian. It did however add a warning and make it not
compile on some systems. It also gave the io_bufsize the value of a
pointer, not the value the user specified.

(this doesn't apply just to Jelmer, I'm just seeing rather too many
checkins from everyone lately where the code could never have possibly
worked, so was obviously checked in without testing)
(This used to be commit 21931ad7a0584c14fd0278adc74eab9cacf14a3d)