Fix references to plain password registry files pointing at an old location.  These files are now in docs/Registry.  For some reason only the PlainPassword files are included in the packaging, not some of the other useful Samba related registry mods such as sign or seal and terminal server.
I also removed the reference to checkinstall as it doesn't seem to
exist on the Solaris system in the build farm and I can't figure out
what it is supposed to do (always a good reason to delete something).

docs.solaris.com says "The checkinstall script is only available with
the Solaris(TM) 2.5 and compatible releases" so maybe this file is
obsolete.

Part of fix for bug 218.
(This used to be commit 0699f362c524dc07b84ad23c57e559ec5e4681f8)