As described in http://davenport.sourceforge.net/ntlm.html add NTLM2 authentication.
NTLM2 is a version of NTLM, that involves both a client and server challenge,
and the creating of a new (presuable more secure) session key.

Unfortunetly this is not quite the same as NTLMv2, and we don't know how to
get the session key.  I suggest looking very closely at what MSCHAPv2, and
other MS auth protocols do...

Andrew Bartlett