This is an intermediate commit ... we check for three different charsets in an iconv_open to be more robust.
However, what we really want to do is find the name for something like CP850,
which can be IBM850 on Solaris, and we might want to also check for
ISO-8859-1.

So, this commit adds a macro to aclocal.m4 that can check for a specific
charset.
(This used to be commit 81dc28ca82b94d1bc42f96273f6cefac6ca1848c)