Follow on from herb's patch for 1.300.2.145 - define SYSV for solaris to fix vendor supplied lp printing (bug #355).
For some reason this is not being set in Samba 3.  In Samba 2.2 the
SYSV define was only set from include/includes.h in a haphazard
fashion.  It's probably better to explicitly define it on a per-operating
system basis anyway.
(This used to be commit e653e13f45c3fa3c2cac792d78e47a2ee0df2b5a)