address bug #359.  Andrew B's patch for implementing client portion of NTLMv2 key exchange.  Also revert the default for 'client ntlmv2 auth' to no.  This caused no ends of grief in different cases.
And based on abartlet's mail....

> All I care about at this point is that we use NTLMv2
> in our client code when connecting to a server that
> supports it.

There is *no* way to tell this.  The server can't tell us, because it
doesn't know what it's DC supports.  The DC can't tell us, because it
doesn't know what the trusted DC supports.  One DC might be Win2k, and
the PDC could be an older NT4.