Fix for bug #299.
There was some confusion over dynamically allocated lists of pointers
(i.e you have to make space for the list of pointers and what they are
pointing too) in the memory buffer passed in from libc.

Valgrind is much happer now and as a bonus there is no segfault.
(This used to be commit 7907c44414acb841a9001e82285790eece73d032)