Changed output so all debug output goes to stderr, and all file processing  goes to stdout.
Note: This change permits use of testparm processing of smb.conf to be
redirected into a file that can be used as an smb.conf file. ie: All
information that should not be in smb.conf will be on stderr, all pertinent
smb.conf info will go to stdout.

Example of use:
	A fully documented smb.conf.master file can be maintained.
	To create smb.conf do:
		testparm -s > smb.conf