First round of merging various UUID structures.
This eliminates RPC_UUID.  It creates the following struct:

struct uuid
{
  uint32 time_low;
  uint16 time_mid;
  uint16 time_hi_and_version;
  uint8  clock_seq[2];
  uint8  node[6];
};
which replaces RPC_UUID and various random struct uuid definitions


and a flat version:
#define UUID_FLAT_SIZE 16
typedef struct uuid_flat
{
	uint8 info[UUID_FLAT_SIZE];
} UUID_FLAT;

which pretty much looks like GUID (which I will start eliminating).

I want us to use the FLAT one only on the wire (perhaps in files, too?), and
I want it to be obvious to the coder that it is the FLAT version.

This leaves a couple of compiler warnings, where GUID isn't completely
replaced by FLAT_UUID yet...I'll get to those soon.