Last of the GUID->struct uuid changes.
GUID has been removed.  If you are dealing with a flattened version of a
[gu]uid, use UUID_FLAT.  smb_pack_uuid and smb_unpack_uuid will switch
between the two.

I'm not sure exactly what the genparse stuff is doing here, so I just
switched it to a UUID_FLAT (no functional change), but I suspect this
may not be the right way to go.
(This used to be commit 001e9d0e814109a26d598001b3c23dfdb04510ff)