Some simple tweaks to the IDL to make it more correct ...
I am also concerned about the difference between NetShareEnum and
NetShareEnumAll. I would be interested to see a trace of NetShareEnumAll.

I also think we should stick to the names used in MSDN where possible, as
that makes it easier for people who are familiar with MSDN to help us get
this right.
(This used to be commit 580939e62e80a06b9a8ad974f6798c49dab4f78c)